requirements.txt
Django==2.1.4
Pillow==5.3.0
pytz==2018.7
sorl-thumbnail==12.5.0
